+++ src/helpers/math.js
@@ -4,6 +4,6 @@ export function modulo(x, max) {
 }
 
 // Test if number `x` is out of `min` ~ `max` range
-export function notInRange(x, min, max) {
-  return x < min || max < x;
+export function inRange(x, min, max) {
+  return min <= x && x <= max;
 }